Hack to enable neovim support
This is a very first attempt at enabling neovim support for
vim-taskwarrir. In my use the only functions that require interaction
are undo and delete, for these two functions there is a check
`has('nvim')` which changes execution from `!` to `terminal`. There is
also no `taskwarrior#refresh()` call, which is done in a BufEnter
autocmd.